requirements/requirements.txt
# Base
pytz==2021.1
python-slugify>=5.0.2
pillow>=7.1.0
# psycopg2==2.8.5 --no-binary psycopg2
psycopg2==2.9.1

# Django
django==3.2.3
# Environment
django-environ==0.4.5
# Crum User
django-crum==0.7.6
# Inputs Forms
django-widget-tweaks==1.4.8
# Export Import
django-import-export==2.3.0
# Merge on Office Open XML
docx-mailmerge==0.5.0
# Django-Cleanup
django-cleanup==5.2.0
# Django REST Framework
#djangorestframework >= 3.9.1
#django-filter==2.0.0
#django-cors-headers >= 3.0.2
django-extensions==3.1.3

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files
#whitenoise==4.1.2

# Celery
redis>=3.2.0
django-redis==4.10.0

celery>=4.3.0
tornado>=4.2.0, <6.0.0
flower==0.9.3

# amqp==5.0.6
# argon2-cffi==20.1.0
# asgiref==3.4.1
# Babel==2.9.1
# billiard==3.6.4.0
# celery==5.1.2
# cffi==1.14.6
# click==7.1.2
# click-didyoumean==0.0.3
# click-plugins==1.1.1
# click-repl==0.2.0
# defusedxml==0.7.1
# diff-match-patch==20200713
# Django==3.2.3
# django-cleanup==5.2.0
# django-crum==0.7.6
# django-environ==0.4.5
# django-extensions==3.1.3
# django-import-export==2.3.0
# django-redis==4.10.0
# django-widget-tweaks==1.4.8
# docx-mailmerge==0.5.0
# et-xmlfile==1.1.0
# flower==0.9.3
# kombu==5.1.0
# lxml==4.6.3
# MarkupPy==1.14
# odfpy==1.4.1
# openpyxl==3.0.7
# Pillow==8.3.1
# prompt-toolkit==3.0.19
# psycopg2==2.9.1
# pycparser==2.20
# PyJWT==1.7.1
# python-slugify==5.0.2
# pytz==2021.1
# PyYAML==5.4.1
# redis==3.5.3
# simplejson==3.17.3
# six==1.16.0
# sqlparse==0.4.1
# tablib==3.0.0
# text-unidecode==1.3
# tornado==5.1.1
# vine==5.0.0
# wcwidth==0.2.5
# xlrd==2.0.1
# xlwt==1.3.0